Refactor server structure by consolidating routes into dedicated controllers for authentication, debugging, domains, and game functionalities. Update route paths and remove deprecated domain routes, enhancing clarity and maintainability.
